refactor: removed useless widgets
