+++ .gitignore
@@ -6,3 +6,4 @@ yonsei_financial_tech_admin/.flutter-plugins-dependencies
 yonsei_financial_tech_admin/.flutter-plugins-dependencies
 yonsei_financial_tech_admin/.flutter-plugins-dependencies
 yonsei_financial_tech_admin/.flutter-plugins-dependencies
+yonsei_financial_tech_admin/.firebase/hosting.YnVpbGQvd2Vi.cache
