--- yonsei_financial_tech_admin/lib/utils/color_constants.dart
@@ -1,5 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ColorConstants {
-  static Color blue = Color(0xFF0D46BB);
-}
